[kernel][vfs] Replace uint32_t with ino_t and dev_t

+++ kernel/include/libkern/libkern.h
@@ -18,13 +18,13 @@ int stoi(void* str, int len);
 void htos(uint32_t hex, char str[]);
 void dtos(uint32_t dec, char str[]);
 void reverse(char s[]);
-uint32_t strlen(const char* s);
+size_t strlen(const char* s);
 int strcmp(const char* a, const char* b);
 int strncmp(const char* a, const char* b, uint32_t num);
-bool str_validate_len(const char* c, uint32_t len);
+bool str_validate_len(const char* c, size_t len);
 
-uint32_t ptrarr_len(const char** s);
-bool ptrarr_validate_len(const char** s, uint32_t len);
+size_t ptrarr_len(const char** s);
+bool ptrarr_validate_len(const char** s, size_t len);
 
 #ifndef max
 #define max(a, b) \
